Use radio buttons instead of checkbox for optional disclosure.
This forces a bit of a change in review output: a checkbox always has a
boolean state, but a checkbox does not (because it might be blank) and
there is not a quick way to convert that to a bool. Even though the
form is completed and filled out, there's no type level guarantee that
the string-based form values are consistent.